🛠[회원가입_약관동의] : 약관동의 컨테이너 생성
